shenations/requirements.txt
asgiref==3.8.1
Django==5.2.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
psycopg2-binary>=2.9
django-cors-headers

python-decouple==3.8
python-dotenv==1.0.1